// src/Types/Statistics.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type StatisticsOut = (Chart | List | Status | Table | Value)[];

export interface Chart {
  type: 'chart';
  payload: {
    type: string;
    field: string;
    data: {
      datasets: ChartDataset[];
      labels: unknown[];
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  metadata?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface ChartDataset {
  data: unknown[];
  label: string;
  [k: string]: unknown;
}
export interface List {
  type: 'list';
  payload: ListPayload;
  metadata?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface ListPayload {
  name: string;
  data: (Status | Value)[];
  [k: string]: unknown;
}
export interface Status {
  type: 'status';
  payload: {
    field: string;
    status: string;
    topic_name: string;
    [k: string]: unknown;
  };
  metadata?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface Value {
  type: 'value';
  payload: {
    type: string;
    field: string;
    topic_name: string;
    unit?: string;
    [k: string]: unknown;
  };
  metadata?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface Table {
  type: 'table';
  payload: TablePayload;
  metadata?: {
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface TablePayload {
  name: string;
  data: ListPayload[];
  [k: string]: unknown;
}